# docs/api.yaml
openapi: 3.0.3
info:
  title: Flask MongoDB API
  description: REST API for managing customers and subscriptions
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.example.com
    description: Production server
paths:
...
  /api/subscriptions:
    get:
      summary: Get all subscriptions
      description: Retrieve a list of subscriptions with optional filtering and pagination
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: status
          in: query
          description: Filter by subscription status
          schema:
            type: string
            enum: [active, canceled, expired, trial]
        - name: customer_id
          in: query
          description: Filter by customer ID
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
      responses:
        "200":
          description: List of subscriptions
          content:
            application/json:
              schema:
                type: object
                properties:
                  subscriptions:
                    type: array
                    items:
                      $ref: "#/components/schemas/Subscription"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "400":
          description: Invalid customer ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      summary: Create a new subscription
      description: Create a new subscription for a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionCreate"
      responses:
        "201":
          description: Subscription created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "400":
          description: Validation error or invalid customer/product ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/subscriptions/{subscription_id}:
    get:
      summary: Get a subscription by ID
      description: Retrieve a specific subscription by its ID
      parameters:
        - name: subscription_id
          in: path
          required: true
          description: Subscription ID
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
      responses:
        "200":
          description: Subscription details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "400":
          description: Invalid subscription ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      summary: Update a subscription
      description: Update an existing subscription
      parameters:
        - name: subscription_id
          in: path
          required: true
          description: Subscription ID
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionUpdate"
      responses:
        "200":
          description: Subscription updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "400":
          description: Validation error or invalid ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "404":
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Delete a subscription
      description: Delete a subscription
      parameters:
        - name: subscription_id
          in: path
          required: true
          description: Subscription ID
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
      responses:
        "200":
          description: Subscription deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Subscription deleted successfully
        "400":
          description: Invalid subscription ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/subscriptions/{subscription_id}/cancel:
    post:
      summary: Cancel a subscription
      description: Cancel an active subscription
      parameters:
        - name: subscription_id
          in: path
          required: true
          description: Subscription ID
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
      responses:
        "200":
          description: Subscription canceled successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "400":
          description: Invalid ID or subscription already canceled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/subscriptions/{subscription_id}/renew:
    post:
      summary: Renew a subscription
      description: Renew a canceled or expired subscription
      parameters:
        - name: subscription_id
          in: path
          required: true
          description: Subscription ID
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
      responses:
        "200":
          description: Subscription renewed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "400":
          description: Invalid ID or subscription cannot be renewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/subscriptions/{subscription_id}/settings:
    get:
      summary: Get subscription settings (GetSettings)
      description: Get effective settings for a subscription (product defaults + custom settings)
      parameters:
        - name: subscription_id
          in: path
          required: true
          description: Subscription ID
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
      responses:
        "200":
          description: Subscription settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  subscription_id:
                    type: string
                    example: "507f1f77bcf86cd799439011"
                  product_defaults:
                    type: object
                    example:
                      {
                        "topBarColor": "#ffffff",
                        "topBarBackgroundColor": "#000000",
                        "positionIndex": 0,
                        "defaultLang": "en",
                      }
                  custom_settings:
                    type: object
                    example: { "topBarColor": "#ff0000", "positionIndex": 2 }
                  effective_settings:
                    type: object
                    example:
                      {
                        "topBarColor": "#ff0000",
                        "topBarBackgroundColor": "#000000",
                        "positionIndex": 2,
                        "defaultLang": "en",
                      }
        "400":
          description: Invalid subscription ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      summary: Update subscription custom settings (EditSubscription)
      description: Update the custom settings for a subscription
      parameters:
        - name: subscription_id
          in: path
          required: true
          description: Subscription ID
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - custom_settings
              properties:
                custom_settings:
                  type: object
                  example:
                    {
                      "topBarColor": "#ff0000",
                      "topBarBackgroundColor": "#00ff00",
                      "positionIndex": 1,
                      "defaultLang": "es",
                    }
      responses:
        "200":
          description: Subscription settings updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "400":
          description: Validation error or invalid ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "404":
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/subscriptions/{subscription_id}/apply-defaults:
    post:
      summary: Apply product defaults to subscription
      description: Apply product default settings to subscription custom_settings (only missing keys)
      parameters:
        - name: subscription_id
          in: path
          required: true
          description: Subscription ID
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
      responses:
        "200":
          description: Product defaults applied successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "400":
          description: Invalid ID or product not customizable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Subscription or product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Subscription:
      type: object
      properties:
        _id:
          type: string
          example: "507f1f77bcf86cd799439011"
        customer_id:
          type: string
          example: "507f1f77bcf86cd799439012"
        product_id:
          type: string
          example: "507f1f77bcf86cd799439013"
        status:
          type: string
          enum: [active, canceled, expired, trial]
          example: "active"
        amount:
          type: number
          example: 29.99
        currency:
          type: string
          example: "USD"
        billing_cycle:
          type: string
          enum: [weekly, monthly, yearly]
          example: "monthly"
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        trial_end_date:
          type: string
          format: date-time
        canceled_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        metadata:
          type: object
          example: { "source": "web", "campaign": "spring2024" }
        custom_settings:
          type: object
          example:
            {
              "topBarColor": "#ff0000",
              "topBarBackgroundColor": "#00ff00",
              "positionIndex": 1,
              "defaultLang": "es",
            }
          description: "Custom settings for subscription customization"
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 10
        total:
          type: integer
          example: 100
        pages:
          type: integer
          example: 10
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Resource not found"
...